refactor: moves field formatting logic to `yara-x-proto` crate.
This change centralizes the logic for handling protobuf field formats, making it accessible to other crates that depend on the proto definitions. It removes duplicated code and ensures a consistent way of handling field formatting across the project.
